requirements.txt
django
djangorestframework
faker
djangorestframework-simplejwt
requests